total rehaul to get it to work on appengine
