+++ app.yaml
@@ -1,5 +1,3 @@
-application: haas-hall-lottery
-version: 1
 runtime: go
 api_version: go1
 
@@ -8,3 +6,6 @@ handlers:
   script: _go_app
 - url: /static
   static_dir: web
+- url: /logo\.png
+  static_files: web/logo.png
+  upload: logo\.png
