+++ main.go
@@ -3,14 +3,9 @@ package lottery
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	//"log"
 	"net/http"
-	//"runtime"
-	"strings"
-	"os"
-	
-//	"google.golang.org/appengine"
+	"strconv"
+	//	"google.golang.org/appengine"
 )
 
 const (
@@ -18,6 +13,129 @@ const (
 	YELLOW = "\u001B[33m"
 	GREEN  = "\u001B[32m"
 	END    = "\u001B[0m"
+	HOME   = `<!doctype html5>
+<html>
+  <head>
+    <title>Haas Hall Academy Lottery</title>
+    <link href="https://www.haashall.org/lottery/style.css" rel="stylesheet" type="text/css" />
+  </head>
+  <body>
+    <main>
+      <form id="container" action="/import" method="post" enctype="multipart/form-data">
+<div id="import">
+  <label for="getData" id="getDataButton">Import CSV Data</label>
+  <input value="Import CSV Data" accept=".csv" type="file" id="getData"/>
+</div>
+<a href="" id="export">export</a>
+<div id="fromServer"></div>
+      </form>
+    </main>
+    <script
+  src="https://code.jquery.com/jquery-3.1.1.min.js"
+  integrity="sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8="
+  crossorigin="anonymous"></script>
+    <script type="text/javascript">
+$(':file').change(function(){
+var file = this.files[0];
+var name = file.name;
+var size = file.size;
+var type = file.type;
+var result;
+      if (file) {
+    var reader = new FileReader();
+    reader.readAsText(file);
+    reader.onload = function(e) {
+    result = e.target.result;
+$.ajax({
+url: 'post',  //Server script to process data
+type: 'POST',
+dataType: 'json',
+        success: completeHandler,
+        data: {'data': result, 'entries': prompt("How many people do you want to admit?")},
+});
+    }
+}
+});
+
+function completeHandler(jqXHR, textStatus) {
+    outputFile = "";
+    $("#fromServer").append(createHeader(jqXHR[0]));
+    jqXHR.forEach(function(entry) {
+    var ret = createElement(entry);
+    var forOutput = ret[0];
+    var forFile = ret[1];
+    $("#fromServer").append(forOutput);
+    $("#export").css("display", "block");
+    outputFile += forFile;
+});
+    $("#getDataButton").css("display","none");
+    var a = document.getElementById("export");
+    var file = new Blob([outputFile], {type: "text/plain"});
+    a.href = URL.createObjectURL(file);
+    a.download = "results.csv";
+}
+
+function createHeader(entry) {
+    output = "<div class=\"entry header\">"
+    for (var key in entry) {
+        if (entry[key] === Object(entry[key])) {
+continue;
+} else {
+    output += "<span class=\"entry-item header\">" + key + "</span>";
+}
+    }
+    return output;
+}
+
+function createElement(entry) {
+    index = 0
+    output = "<div class=\"entry\">"
+    fileData = ""
+    for (var key in entry) {
+if (entry.hasOwnProperty(key)) {
+    if (entry[key] === Object(entry[key])) { 
+continue;
+    } else { // this is for the real information
+if (index % 2 == 0) { // even
+    fileData += entry[key] + ", ";
+    output += "<span class=\"entry-item even\">" + formatAdmittance(key, entry[key]) + "</span>";
+} else {
+    fileData += entry[key] + ", ";
+    output += "<span class=\"entry-item odd\">" + formatAdmittance(key, entry[key]) + "</span>";
+}
+index++
+    }
+}
+    }
+    fileData += "\n";
+    output += "</div>\n"
+    return [output, fileData];
+}
+
+function formatAdmittance(key, info) {
+    if (key !== "Status" && key !== "Priority") {
+       return info;
+    }
+    if (key === "Status") {
+if (info == 2) {
+    return "<span id=\"admitted\">Admitted</span>";
+} else if (info == 1) {
+    return "<span id=\"waitlisted\">Waitlisted</span>";
+}
+    }
+    if (key === "Priority") {
+if (info == 2) {
+    return "<span id=\"faculty\">Faculty</span>";
+} else if (info == 1) {
+    return "<span id=\"sibling\">Sibling</span>";
+} else {
+    return "";
+}
+    }
+}
+    </script>
+  </body>
+</html>`
 )
 
 type fileError struct {
@@ -38,41 +156,22 @@ func formatPrint(s string, color string) string {
 // Writes a 404 message to the page
 func NotFound(w http.ResponseWriter, r *http.Request) {
 	fmt.Println(formatPrint("REQUEST[404]: ", RED), r.URL)
-	s, _ := OpenPage("404.html")
 	w.Header().Set("Content-Type", "text/html")
-	w.Write([]byte(s))
+	w.Write([]byte("nothing to see here"))
 	w.WriteHeader(http.StatusNotFound)
-}
-
-// Opens a web page in the /static/ directory and returns as a string
-// As we aren't dealing with *massive* files, and since this is self-serving
-// we don't have to worry about network security (since the program doesn't interact with
-// a network).
-func OpenPage(fileName string) (string, error) {
-	wd, err := os.Getwd()
-	if err != nil {
-		panic(err)
-	}
-	data, err := ioutil.ReadFile(wd + "/web/" + fileName)
-	if err != nil {
-		return "", fileError{"File not found."}
-	}
-	return string(data), nil
 }
 
 // Main function to serve the site
 func init() {
 	// serves for the webpage "/" — i.e. the homepage
-	
+
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
 		if r.URL.Path != "/" { // we want "/" with no appending text
 			NotFound(w, r)
 			return
 		}
-		s, _ := OpenPage("home.html") // the home page content
 		w.Header().Set("Content-Type", "text/html")
-		w.Write([]byte(s))                                       // prints to webpage
-		fmt.Println(formatPrint("REQUEST[200]: ", GREEN), r.URL) // prints to console
+		w.Write([]byte(HOME))                                    // prints to webpage
 	})
 
 	http.HandleFunc("/post", func(w http.ResponseWriter, r *http.Request) {
@@ -84,35 +183,15 @@ func init() {
 		entries := CSVtoEntries(r.FormValue("data"))
 		sorted := Sort(entries)
 		prioritized := Prioritize(sorted)
-		admitted := Admit(prioritized, 3)
+		i, err := strconv.Atoi(r.FormValue("entries"))
+		if err != nil {
+			return
+		}
+		admitted := Admit(prioritized, i)
 		data, err := json.Marshal(admitted) // the error needs to do something
 		if err != nil {
 			fmt.Println(err)
 		}
 		w.Write(data)
-		fmt.Println(formatPrint("REQUEST[200]: ", GREEN), r.URL) // prints to console
 	})
-
-	http.HandleFunc("/static/", func(w http.ResponseWriter, r *http.Request) {
-		if !strings.HasPrefix(r.URL.Path, "/static") {
-			NotFound(w, r)
-		}
-		s, err := OpenPage(r.URL.Path[1:])
-		if err != nil {
-			NotFound(w, r)
-			return
-		}
-		if strings.HasSuffix(r.URL.Path, ".jpg") {
-			w.Header().Set("Content-Type", "image/jpg")
-		} else if strings.HasSuffix(r.URL.Path, ".html") {
-			w.Header().Set("Content-Type", "text/html")
-		} else if strings.HasSuffix(r.URL.Path, ".css") {
-			w.Header().Set("Content-Type", "text/css")
-		} else if strings.HasSuffix(r.URL.Path, ".js") {
-			w.Header().Set("Content-Type", "text/javascript")
-		}
-		w.Write([]byte(s))
-		fmt.Println(formatPrint("REQUEST[200]: ", YELLOW), r.URL)
-	})
-
 }
