--- web/static/main.css
@@ -1,139 +1,0 @@
-body {
-     margin: 0px;
-     padding: 0px;
-     background: #fff;
-}
-
-
-hr {
-   color: #ffc63d;
-   border-color: #ffc63d;
-   background-color: #ffc63d;
-   margin: 0px 20px;
-}
-
-main {
-     position: relative;
-     width: 100%;
-     height: 100%;
-     background: #005198;
-     float: right;
-     padding-top: 3.5%;
-     z-index: 3;
-}
-
-main:after {
-    content: ' ';
-    display: block;
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    z-index: -1;
-    opacity: 0.3;
-    background-image: url('/static/bluebackground.jpg');
-    background-repeat: no-repeat;
-    background position: 50% 0;
-    -ms-background-size: cover;
-    -o-background-size: cover;
-    -moz-background-size: cover;
-    -webkit-background-size: cover;
-    background-size: cover;
-}
-
-#container {
-    display: table;
-    background: #eee;
-    height: 90%;
-    width: 97%;
-    margin: 15px;
-    z-index: -1;
-    text-align: center;
-}
-
-#import {
-    line-height: 100%;
-    display: table-cell;
-    text-align: center;
-    vertical-align: middle;
-}
-
-#import label {
-    background: rgb(0, 81, 152);
-    color: white;
-    font-size: 18px;
-    font-family: helvetica;
-    padding: 5px;
-    border-radius: 2px;
-}
-
-#import label:hover {
-	border: 1px solid black;
-}
-
-#import label:active {
-	font-weight: bold;
-}
-
-#import input {
-	display: none;
-}
-
-#fromServer {
-	    display: table;
-	    width: 100%;
-	    background-color: #eee;
-	    text-align: center;
-}
-
-#export {
-    display: none;
-    margin: 10px;
-    background: #0560b7;
-    color: white;
-    font-weight: bold;
-    text-decoration: none;
-    /* border: 3px solid yellow; */
-    /* border-radius: 100px; */
-    font-family: sans-serif;
-    padding: 5px;
-}
-
-.entry {
-	  display: table-row;
-}
-
-.entry-item {
-        display: table-cell;
-        width: 200px;
-}
-
-.header {
-	background: #555;
-	color: white;
-}
-
-#admitted {
-	  color: green;
-}
-
-#waitlisted {
-	  color: blue;
-}
-
-#faculty {
-	 color: red;
-}
-
-#sibling {
-	 color: purple;
-}
-
-.odd {
-        background: #d4d4d4;
-}
-
-.even {
-       background: #ccc;
-}
